refactor: move mergeFacets to the app directory
